split App.tsx into utils.ts, summary.ts and Viz.tsx
